Removed addShiftViewController and descriptionViewController.
